+++ packages/databricks-vscode/src/configuration/LoginWizard.ts
@@ -304,12 +304,16 @@ export class LoginWizard {
             | PersonalAccessTokenAuthProvider;
         switch (pick.authType) {
             case "azure-cli":
-                authProvider = new AzureCliAuthProvider(this.state.host!);
+                authProvider = new AzureCliAuthProvider(
+                    this.state.host!,
+                    this.cliWrapper
+                );
                 break;
 
             case "databricks-cli":
                 authProvider = new DatabricksCliAuthProvider(
                     this.state.host!,
+                    this.cliWrapper.cliPath,
                     this.cliWrapper
                 );
                 break;
@@ -329,7 +333,8 @@ export class LoginWizard {
                     }
                     authProvider = new PersonalAccessTokenAuthProvider(
                         this.state.host!,
-                        token
+                        token,
+                        this.cliWrapper
                     );
                 }
                 break;
@@ -435,7 +440,7 @@ export async function listProfiles(cliWrapper: CliWrapper) {
         async () => {
             const profiles = (
                 await cliWrapper.listProfiles(
-                    workspaceConfigs.databrickscfgLocation
+                    FileUtils.getDatabricksConfigFilePath().fsPath
                 )
             ).filter((profile) => {
                 try {
@@ -448,6 +453,12 @@ export async function listProfiles(cliWrapper: CliWrapper) {
 
             return profiles;
         }
+    );
+}
+
+export async function getProfilesForHost(host: URL, cliWrapper: CliWrapper) {
+    return (await listProfiles(cliWrapper)).filter(
+        (profile) => profile.host?.toString() === host.toString()
     );
 }